Fix #750: Implement abstract class BaseCurl for Curl and MultiCurl
